Move value.hpp in value directory
