+++ include/parser.hpp
@@ -6,7 +6,7 @@
 #include <vector>
 
 #include "string_utils.hpp"
-#include "value.hpp"
+#include "value/value.hpp"
 
 namespace tkg
 {
